Makefile.am: Under distcheck, specify plugin dir
Without this, when distcheck runs, it does:

 * make
 * make check
 * make install

... and the make check phase runs saslpasswd2 which requires
plugins to be in the plugindir. But without `make install` being
run, they aren't there, so we have tell configure to look elsewhere
when under distcheck.

This is a hack. The real solution is get `make check` working against
a non-installed build, somehow...